Completed Combat System
Moving/Firing is done solely with Right Click; UI buttons are not implemented.

ActionPanel now refers to the Player, rather than the other way around.

Moving takes a destination tile as a parameter. Added a 5sec failsafe to revert to Idle in the event Moving becomes stuck.

Known Issues:
A dead character's Tile does not become available until being walked over.

Idle checks for a valid Tile before acknowledging the Move. Moving also uses the ProcessAction function.  This leads to a double "Not a valid tile" log but is otherwise fine. We may need to have ProcessAction return/out a bool. "If (ProcessAction.Move == true) ChangeState(Moving)" sort of thing.