Functionality implemented.
Player can go through the Choose/Use item progression.

Items are prefabs. Inventory script makes a duplicate list of this.
Selecting ChooseItem brings up the character's inventory. Selecting an item moves to UseItem where you can select/tab through the appropriate targets.  There are some inconsistencies with moving backward, but I think those are already fixed in Master, so I'm waiting til Merge to check that.

Items have Charges for us to play with. For instance, Medkits have 1 Charge and are removed once depleted. However, rather than having 3 grenades take up slots, we can also set it to one slot has 3 grenades to use. We can toy with this.

There is no current animation/feedback other than Debug.Log.

Depleting an item's charges removes it from the List. This automatically reorganizes the list ( Item[1] moves to Item[0] if the slot is empty). I'm leaving this as is for now, as it should be fine. But if we want to be able to reorganize items, that's another effort.