PlayerTurnOverState implemented
End Turn button can now only be pressed when all good characters are out of AP. This was done by implementing a new CombatState: PlayerTurnOverState, which is only active in case it's still the player turn but no characters are able to do anything, in which case the End Turn button is usable.  AP availability is checked on Update.