Finished implementing end turn button
Now the end turn button will not be clickable if any player unit still has AP left. If it is clicked while a player unit still has AP, the unit with the most AP left will be selected.
When the state change from player turn to enemy turn and vice versa, a debug log shows which turn has started.