Added unitType as a stat of Unit
Units now have a stat called unitType, and enum which is currently either HUMAN or ROBOTIC. Easily expandable in the future.
UseItem now returns a bool value specifying whether or not the item was used successfully. If the item is not compatible with the unitType of the target, the function will return false and no AP will be used.