Simplified camera movement parameters.
*Pics posted in Discord*

Using the Frustrum wouldn't work (move the camera) after rotating the camera. I instead replaced it to check for a point on the Top 3rd on the screen. If that point contains a GridMap tile, the camera may move.

It sometimes gets stuck on parts of the level with empty or no tiles, but that can be fixed in level design.

The original frustrum controls are commented out in case someone with more experience wants to do a deeper dive.